[Refactor + Style] (code) Move the examples code to './tests/' and modify importing.
