+++ tests/unit_test/adapter/lock.py
@@ -1,15 +1,15 @@
-from multirunnable.mode import RunningMode, FeatureMode
-from multirunnable.adapter import Lock, RLock, Semaphore, BoundedSemaphore
-from multirunnable.adapter.lock import AsyncLock, AsyncSemaphore, AsyncBoundedSemaphore
-
-from ..test_config import Semaphore_Value
-from .._examples import RunByStrategy
-from ..framework.lock import LockTestSpec, RLockTestSpec, SemaphoreTestSpec, BoundedSemaphoreTestSpec
-
 from typing import Union
 # import asyncio
 import pytest
 import sys
+
+from multirunnable.adapter.lock import AsyncLock, AsyncSemaphore, AsyncBoundedSemaphore
+from multirunnable.adapter import Lock, RLock, Semaphore, BoundedSemaphore
+from multirunnable.mode import RunningMode, FeatureMode
+
+from ...test_config import Semaphore_Value
+from ..._examples import RunByStrategy
+from ..framework.lock import LockTestSpec, RLockTestSpec, SemaphoreTestSpec, BoundedSemaphoreTestSpec
 
 
 _Semaphore_Value: int = Semaphore_Value
