+++ tests/unit_test/api/decorator.py
@@ -1,17 +1,25 @@
-from multirunnable import PYTHON_MAJOR_VERSION, PYTHON_MINOR_VERSION
-from multirunnable.mode import RunningMode
+from gevent.threading import get_ident as get_gevent_ident
+from gevent import sleep as gevent_sleep
+import threading
+import asyncio
+import pytest
+import time
+import os
+
+from multirunnable.coroutine.strategy import AsynchronousStrategy
+from multirunnable.parallel.share import Global_Manager
 from multirunnable.api.decorator import RunWith, AsyncRunWith
 from multirunnable.factory.lock import LockFactory, SemaphoreFactory, BoundedSemaphoreFactory
-from multirunnable.parallel.share import Global_Manager
-from multirunnable.coroutine.strategy import AsynchronousStrategy
+from multirunnable.mode import RunningMode
+from multirunnable import PYTHON_MAJOR_VERSION, PYTHON_MINOR_VERSION
 
-from ..framework.lock import LockTestSpec, SemaphoreTestSpec
-from ..test_config import Worker_Size, Worker_Pool_Size, Task_Size, Semaphore_Value
-from .._examples import RunByStrategy
-from .._examples_with_synchronization import (
+from ...test_config import Worker_Size, Worker_Pool_Size, Task_Size, Semaphore_Value
+from ..._examples import RunByStrategy
+from ..._examples_with_synchronization import (
     instantiate_lock, instantiate_rlock,
     instantiate_semaphore, instantiate_bounded_semaphore
 )
+from ..framework.lock import LockTestSpec, SemaphoreTestSpec
 from ._retry_sample import (
     _Retry_Time, _Default_Retry_Time, _Test_Return_Value, _Test_Exception,
     init_flag, get_process_flag,
@@ -21,14 +29,6 @@ from ._retry_sample import (
     async_target_function, async_target_function_raising_exception,
     TargetBoundedFunction, TargetBoundedAsyncFunction
 )
-
-from gevent.threading import get_ident as get_gevent_ident
-from gevent import sleep as gevent_sleep
-import threading
-import asyncio
-import pytest
-import time
-import os
 
 
 _Worker_Size = Worker_Size
