+++ tests/unit_test/concurrent/queue.py
@@ -1,7 +1,7 @@
 # from multirunnable.concurrent.queue import ThreadQueueType
 from multirunnable import PYTHON_MAJOR_VERSION, PYTHON_MINOR_VERSION
 
-from ..test_config import Semaphore_Value
+from ...test_config import Semaphore_Value
 
 if PYTHON_MAJOR_VERSION == 3 and PYTHON_MINOR_VERSION == 6:
     from queue import Queue, LifoQueue, PriorityQueue
