+++ tests/unit_test/coroutine/queue.py
@@ -1,6 +1,6 @@
 from multirunnable import PYTHON_MAJOR_VERSION, PYTHON_MINOR_VERSION
 
-from ..test_config import Semaphore_Value
+from ...test_config import Semaphore_Value
 
 if PYTHON_MAJOR_VERSION == 3 and PYTHON_MINOR_VERSION == 6:
     from gevent.queue import (
