+++ tests/unit_test/executor.py
@@ -1,13 +1,16 @@
-from multirunnable import get_current_mode, set_mode, RunningMode, SimpleExecutor, PYTHON_MAJOR_VERSION, PYTHON_MINOR_VERSION
-from multirunnable.executor import AdapterExecutor
-from multirunnable.parallel.strategy import ProcessStrategy
-from multirunnable.parallel.share import Global_Manager
+import threading
+import pytest
+import time
+
 from multirunnable.concurrent.strategy import ThreadStrategy
 from multirunnable.coroutine.strategy import GreenThreadStrategy, AsynchronousStrategy
+from multirunnable.parallel.strategy import ProcessStrategy
+from multirunnable.parallel.share import Global_Manager
+from multirunnable.executor import AdapterExecutor
+from multirunnable import get_current_mode, set_mode, RunningMode, SimpleExecutor
 
-from .framework.strategy import GeneralRunningTestSpec
-from .test_config import Worker_Size, Running_Diff_Time, Test_Function_Sleep_Time, Test_Function_Args
-from ._examples import (
+from ..test_config import Worker_Size, Running_Diff_Time, Test_Function_Sleep_Time, Test_Function_Args
+from .._examples import (
     # # Import the falgs
     get_running_cnt, get_current_workers, get_running_workers_ids, get_running_done_timestamps,
     # # Import some common functions
@@ -15,10 +18,7 @@ from ._examples import (
     # # Import some target functions to run for Pool object
     target_function, target_function_for_map
 )
-
-import threading
-import pytest
-import time
+from .framework.strategy import GeneralRunningTestSpec
 
 
 _Worker_Size = Worker_Size
