+++ tests/unit_test/framework/lock.py
@@ -1,19 +1,19 @@
-from multirunnable import PYTHON_MAJOR_VERSION, PYTHON_MINOR_VERSION
-from multirunnable.mode import FeatureMode
-from multirunnable.adapter import Lock, RLock, Semaphore, BoundedSemaphore
-
-from ..test_config import Worker_Size, Worker_Pool_Size, Task_Size, Semaphore_Value
-
-from abc import ABCMeta, abstractmethod, ABC
-from typing import Union
 from gevent.threading import get_ident as get_gevent_ident
 from gevent import sleep as gevent_sleep
+from typing import Union
+from abc import ABCMeta, abstractmethod, ABC
 import multiprocessing
 import threading
 import asyncio
 import pytest
 import random
 import time
+
+from multirunnable.adapter import Lock, RLock, Semaphore, BoundedSemaphore
+from multirunnable.mode import FeatureMode
+from multirunnable import PYTHON_MAJOR_VERSION, PYTHON_MINOR_VERSION
+
+from ...test_config import Worker_Size, Worker_Pool_Size, Task_Size, Semaphore_Value
 
 
 _Worker_Size = Worker_Size
